+++ core/src/main/java/arez/RunProcedureAsActionReaction.java
@@ -43,6 +43,7 @@ final class RunProcedureAsActionReaction
     }
     observer.getContext().action( Arez.areNamesEnabled() ? observer.getName() : null,
                                   Arez.shouldEnforceTransactionType() ? observer.getMode() : null,
+                                  false,
                                   action,
                                   true,
                                   observer );